# api/openapi.yaml
openapi: 3.1.0
info:
  title: Afterglow Reconciliation API
  version: 1.0.0
  description: Synthetic DOOH reservation and playback contract. Not a Vistar API implementation. Money is integer USD micros. Tenant comes from the deployment credential. Demo mode is loopback-only and bypasses bearer authentication.
servers:
  - url: http://127.0.0.1:8090
security:
  - bearer: []
paths:
  /api/v1/deliveries/{id}:
    get:
      operationId: getDeliveryEvidence
      summary: Read an exact receipt and its reservation from a consistent database snapshot
      description: Tenant-scoped evidence independent of the recent snapshot window. Outbox acknowledgement may lag a committed settlement. A receipt for an unknown reservation omits reservation.
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string, minLength: 1, maxLength: 100, pattern: '^[a-zA-Z0-9_-]+$'}
      responses:
        '200': {description: 'DeliveryEvidence with delivery, optional reservation, outbox_state, dispatch_attempts and observed_at in Unix milliseconds.'}
        '401': {description: Invalid credential}
        '404': {description: Receipt not found in the authenticated tenant}
        '422': {description: Invalid identifier}
        '503': {description: Storage unavailable}
  /api/v1/airports:
    get:
      operationId: searchAirportCatalog
      summary: Search real US scheduled-service airport reference data
      description: Proxies the separate Go airport service. Includes US territories. No advertising rights or real-time flight information is implied. Each response includes source, scope, fetch time, version hash and stale status.
      parameters:
        - {in: query, name: q, schema: {type: string, maxLength: 100}}
        - {in: query, name: region, schema: {type: string, maxLength: 20}}
        - {in: query, name: limit, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
        - {in: query, name: offset, schema: {type: integer, minimum: 0, default: 0}}
      responses:
        '200': {description: "Current page in airports, all matching locations in map_airports, matched/total counts, regions, fetched_at, source_modified, version, source, scope, stale and warning. Sorted by source identifier."}
        '400': {description: "Invalid filters or pagination"}
        '401': {description: "Invalid credential"}
        '503': {description: "Catalog has not loaded or the separate service is unavailable. Financial APIs continue independently."}
  /api/v1/receipts/batch:
    post:
      operationId: acceptReceiptBatch
      summary: Accept up to 50 receipts with independent transaction outcomes
      description: Maximum request body 32768 bytes. Items consume individual rate-limit tokens. Retry ambiguous or transient failures using the same event IDs and payloads. Unsupported schema versions can be durably accepted and later quarantined by the worker. This is a synthetic partner contract.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [receipts]
              properties:
                receipts:
                  type: array
                  minItems: 1
                  maxItems: 50
                  items: {$ref: '#/components/schemas/Receipt'}
      responses:
        '202': {description: "All items durably accepted or replayed. Body includes accepted/rejected/replayed counts and ordered results with zero-based index, status, id, replayed, retryable and detail."}
        '207': {description: "One or more items rejected. Inspect each result; 400 malformed item, 422 invalid identifiers, 409 conflicting event identity, 503 transient storage error. Accepted items remain committed."}
        '400': {description: "Malformed, oversized or unknown-field envelope; no items attempted"}
        '401': {description: "Invalid credential"}
        '415': {description: "JSON content type required"}
        '422': {description: "Batch must contain 1 to 50 items; no items attempted"}
        '429': {description: "Insufficient per-item admission tokens; no items attempted; Retry-After provided"}
        '503': {description: "Request concurrency capacity unavailable"}
  /healthz:
    get:
      operationId: getLiveness
      security: []
      summary: Process liveness
      responses:
        '200': {description: "Process is alive"}
  /readyz:
    get:
      operationId: getReadiness
      security: []
      summary: Database connectivity and shutdown readiness
      responses:
        '200': {description: "Accepting work"}
        '503': {description: "Not ready"}
  /api/v1/reservations:
    post:
      operationId: reservePlayback
      summary: Hold budget for one screen playback
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: {type: string, pattern: '^[A-Za-z0-9_-]{1,100}$'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [campaign_id, screen_id, cost_micros]
              properties:
                campaign_id: {type: string}
                screen_id: {type: string}
                cost_micros: {type: integer, minimum: 1, maximum: 1000000000}
      responses:
        '201':
          description: Budget held. Reservation play window lasts 15 minutes.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Reservation'}
        '200': {description: "Existing reservation replayed for the same normalized payload and key. Includes Idempotent-Replayed header."}
        '409': {description: "Insufficient remaining budget or idempotency key conflicts with prior content"}
        '404': {description: "Unknown campaign or screen"}
        '422': {description: "Invalid IDs or cost"}
        '400': {description: "Malformed or unknown JSON fields"}
        '401': {description: "Invalid credential"}
        '415': {description: "JSON content type required"}
        '429': {description: "Local mutation rate limit reached; Retry-After provided"}
        '503': {description: "Capacity or dependency unavailable"}
  /api/v1/receipts:
    post:
      operationId: acceptPlaybackReceipt
      summary: Persist a playback receipt and dispatch intent atomically
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Receipt'}
      responses:
        '202':
          description: Receipt and outbox committed. Not a promise of broker delivery or settlement. Event ID replay returns the existing delivery ID.
          content:
            application/json:
              schema:
                type: object
                required: [id, replayed, durability]
                properties:
                  id: {type: string}
                  replayed: {type: boolean}
                  durability: {const: database_and_outbox_committed}
        '409': {description: "Event ID reused with different normalized content"}
        '422': {description: "Invalid identifiers"}
        '400': {description: "Invalid or oversized JSON"}
        '401': {description: "Invalid credential"}
        '415': {description: "JSON content type required"}
        '429': {description: "Rate limited"}
        '503': {description: "Capacity or database unavailable"}
  /api/v1/deliveries/{id}/replay:
    post:
      operationId: replayDelivery
      summary: Requeue immutable evidence after investigating the cause
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        '202': {description: "Quarantined or failed receipt requeued; normal validation still applies"}
        '409': {description: "Delivery not eligible for replay or not visible to this tenant"}
        '401': {description: "Invalid credential"}
        '503': {description: "Database unavailable"}
  /api/v1/snapshot:
    get:
      operationId: getOperationsSnapshot
      summary: Consistent tenant-scoped operations snapshot
      responses:
        '200':
          description: Campaign balances, shared screen inventory, newest 100 reservations and receipts, newest 80 audit records, all-time status counts and oldest_pending_ms. Timestamps are Unix milliseconds. PostgreSQL reads use a repeatable-read transaction.
        '401': {description: "Invalid credential"}
        '503': {description: "Database unavailable"}
  /api/v1/runtime:
    get:
      operationId: getRuntime
      summary: Execution mode, storage, role, paused state and breaker state
      responses:
        '200': {description: "Runtime configuration without credentials"}
        '401': {description: "Invalid credential"}
  /api/v1/metrics:
    get:
      operationId: getPrometheusMetrics
      summary: Prometheus counters and processing duration
      responses:
        '200': {description: "Prometheus exposition format"}
        '401': {description: "Invalid credential"}
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  schemas:
    Reservation:
      type: object
      properties:
        id: {type: string}
        campaign_id: {type: string}
        screen_id: {type: string}
        cost_micros: {type: integer}
        state: {type: string, enum: [held, settled, released]}
        created_at: {type: integer, description: Unix milliseconds}
        expires_at: {type: integer, description: End of play window; funds released after additional 15-minute receipt grace}
    Receipt:
      type: object
      additionalProperties: false
      required: [event_id, reservation_id, screen_id]
      description: Only identifiers are synchronously validated. Unsupported version, missing or invalid play time or duration are durably accepted and then quarantined by the consumer. This separation preserves partner evidence.
      properties:
        schema_version: {type: integer, examples: [1]}
        event_id: {type: string, pattern: '^[A-Za-z0-9_-]{1,100}$'}
        reservation_id: {type: string, pattern: '^[A-Za-z0-9_-]{1,100}$'}
        screen_id: {type: string, pattern: '^[A-Za-z0-9_-]{1,100}$'}
        played_at: {type: integer, description: Unix milliseconds; must fall inside reservation window}
        duration_ms: {type: integer, description: Consumer accepts 1000 through 120000 inclusive}